Scala 3 code compiles (but with warnings: will fix)
A lot of TODOs left.
