+++ scala3/src/main/scala/math/cat/topos/CategoryOfDiagrams.scala
@@ -2,13 +2,14 @@ package math.cat.topos
 
 import scala.language.implicitConversions
 import scala.language.postfixOps
-
+import reflect.Selectable.reflectiveSelectable
 import math.Base._
-import math.cat._
-import math.cat.topos.CategoryOfDiagrams.{BaseCategory, _}
+import math.cat.{topos, _}
+import math.cat.topos.CategoryOfDiagrams.BaseCategory
 import math.sets.Sets._
 import math.sets._
 import scalakittens.Result
+import topos.CategoryOfDiagrams._
 
 class CategoryOfDiagrams(val domain: Category)
   extends Category
@@ -19,15 +20,6 @@ class CategoryOfDiagrams(val domain: Category)
   type Node = Diagram
   override type Obj = Diagram
   override type Arrow = DiagramArrow
-  
-  private def constSet(name: String)(obj: BaseCategory.Obj): Diagram =
-    const(name, topos)(obj.asInstanceOf[set])
-  
-  override lazy val initial: Result[Obj] = BaseCategory.initial map constSet("initial")
-  lazy val _0: Obj = initial iHope
-  
-  override lazy val terminal: Result[Obj] = BaseCategory.terminal map constSet("terminal")
-  lazy val _1: Obj = terminal iHope
   
   lazy val subterminals: Set[Diagram] = {
     def objectMapping(candidate: Set[domain.Obj]) =
@@ -128,85 +120,7 @@ class CategoryOfDiagrams(val domain: Category)
     probablyFunctor.iHope
   }
 
-  private[topos] case class product2builder(x: Diagram, y: Diagram) {
-
-    private def productAt(o: domain.Obj) = Sets.product2(x(o), y(o))
-    private def mappingOfObjects(o: domain.Obj): set = productAt(o).untyped
-    
-    def transition(z: Diagram)(a: domain.Arrow)(pz: Any) = {
-      val sf: SetFunction = z.asFunction(z.arrowsMapping(z.d0.arrow(a)))
-      sf(pz)
-    }
-    
-    private def mappingOfArrows(a: domain.Arrow): SetFunction = {
-      val from = productAt(domain.d0(a))
-      val to = productAt(domain.d1(a))
-      def f(p: Any): Any = p match {
-        case (px, py) => (transition(x)(a)(px), transition(y)(a)(py))
-        case other =>
-          throw new IllegalArgumentException(s"Expected a pair of values, got $other")
-      }
-      new SetFunction("", from.untyped, to.untyped, f)
-    }
-    
-    val diagram = Diagram(s"${x.tag}×${y.tag}", topos)(mappingOfObjects, a => mappingOfArrows(a))
-  }
-
-  /**
-    * Cartesian product of two diagrams
-    * TODO: figure out how to ensure the same d0 in both Di
-    */
-  def product2(x: Diagram, y: Diagram): Diagram = product2builder(x, y).diagram
-
-  /**
-    * Given arrows `f` and `g`, builds an arrow (f×g): dom(f)×dom(g) -> codom(f)×codom(g)
-    *
-    * @param f first component
-    * @param g second component
-    * @return a product of `f` and `g`
-    */
-  def productOfArrows(f: DiagramArrow, g: DiagramArrow): DiagramArrow = {
-
-    val mapping: Mapping = x => {
-      val fx = f(x).asInstanceOf[SetMorphism[Any, Any]]
-      val gx = g(x).asInstanceOf[SetMorphism[Any, Any]]
-      
-      { case (a, b) => (fx(a), gx(b)) }
-    }
-
-    buildArrow(
-      concat(f.tag, "×", g.tag),
-      product2(f.d0, g.d0),
-      product2(f.d1, g.d1),
-      mapping)
-  }
-
   def inclusionOf(p: Point): Includer = inclusionOf(p.asDiagram)
-  
-  def standardInclusion(p: Point, d: Diagram): Result[Arrow] = {
-    inclusionOf(p) in d map { q => {
-      uniqueFromTerminalTo(p) andThen q named p.tag
-    } }
-  }
-
-  /**
-    * An arrow from terminal to the point as a diagram
-    * @return
-    */
-  def uniqueFromTerminalTo(p: Point): Arrow = {
-    new DiagramArrow {
-      val tag = p.tag
-
-      override val d0: Diagram = _1
-      override val d1: Diagram = p.asDiagram
-
-      override def transformPerObject(o: domainCategory.Obj): codomainCategory.Arrow =
-      codomainCategory.arrow {
-        val value = p(o)
-        new SetFunction(s"tag($o)", _1(o), Set(value), _ => value)
-      }
-    }    
-  }
 }
 
 object CategoryOfDiagrams {
