+++ scala3/src/main/scala/math/cat/topos/Topos.scala
@@ -5,13 +5,15 @@ import math.cat.Category
 /**
   * Topos functionality
   */
-trait Topos[O, A] { self: Category =>
+trait Topos[O, A] {
   type Obj = O
   type Arrow = A
   // finite limits should exist
   def product2(x: O, y: O): O
   
   val Ω: O
+
+  def ΩxΩ: O
   
 //  def pow: Obj => Obj
 }
